+++ FbonizziMonoGameAndroid/FbonizziMonoGameAndroid/AndroidSettingsRepository.cs
@@ -1,14 +1,21 @@
-﻿namespace FbonizziMonoGameAndroid
-{
-    using Android.Content;
-    using Android.Preferences;
-    using FbonizziMonogame;
-    using System;
+﻿using Android.Content;
+using Android.Preferences;
+using FbonizziMonoGame.PlatformAbstractions;
+using System;
 
+namespace FbonizziMonoGameAndroid
+{
+    /// <summary>
+    /// Android Xamarin implementation of <see cref="ISettingsRepository"/>
+    /// </summary>
     public class AndroidSettingsRepository : ISettingsRepository
     {
         private readonly ISharedPreferences _settings;
 
+        /// <summary>
+        /// Android Xamarin implementation of <see cref="ISettingsRepository"/>
+        /// </summary>
+        /// <param name="context"></param>
         public AndroidSettingsRepository(Context context)
         {
             if (context == null)
@@ -17,6 +24,12 @@
             _settings = PreferenceManager.GetDefaultSharedPreferences(context);
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public bool GetOrSetBool(string key, bool defaultValue)
         {
             if (_settings.All.ContainsKey(key))
@@ -26,6 +39,12 @@
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public int GetOrSetInt(string key, int defaultValue)
         {
             if (_settings.All.ContainsKey(key))
@@ -35,6 +54,12 @@
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public long GetOrSetLong(string key, long defaultValue)
         {
             if (_settings.All.ContainsKey(key))
@@ -44,6 +69,12 @@
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public string GetOrSetString(string key, string defaultValue)
         {
             if (_settings.All.ContainsKey(key))
@@ -53,6 +84,12 @@
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public TimeSpan GetOrSetTimeSpan(string key, TimeSpan defaultValue)
         {
             long ticksValue = defaultValue.Ticks;
@@ -64,6 +101,12 @@
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public DateTime GetOrSetDateTime(string key, DateTime defaultValue)
         {
             long ticksValue = defaultValue.ToBinary();
@@ -75,6 +118,11 @@
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetInt(string key, int value)
         {
             var editor = _settings.Edit();
@@ -82,6 +130,11 @@
             editor.Commit();
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetBool(string key, bool value)
         {
             var editor = _settings.Edit();
@@ -89,6 +142,11 @@
             editor.Commit();
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetLong(string key, long value)
         {
             var editor = _settings.Edit();
@@ -96,6 +154,11 @@
             editor.Commit();
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetString(string key, string value)
         {
             var editor = _settings.Edit();
@@ -103,12 +166,22 @@
             editor.Commit();
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetTimeSpan(string key, TimeSpan value)
         {
             long ticksValue = value.Ticks;
             SetLong(key, ticksValue);
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetDateTime(string key, DateTime value)
         {
             long ticksValue = value.ToBinary();
