+++ FbonizziMonoGameAndroid/FbonizziMonoGameAndroid/AndroidTextFileImporter.cs
@@ -1,18 +1,31 @@
-﻿namespace FbonizziMonoGameAndroid
-{
-    using Android.Content.Res;
-    using FbonizziMonogame.Assets;
-    using System.IO;
+﻿using Android.Content.Res;
+using FbonizziMonoGame.PlatformAbstractions;
+using System;
+using System.IO;
 
+namespace FbonizziMonoGameAndroid
+{
+    /// <summary>
+    /// Android Xamarin implementation of <see cref="ITextFileLoader"/>
+    /// </summary>
     public class AndroidTextFileImporter : ITextFileLoader
     {
-        private AssetManager _assets;
+        private readonly AssetManager _assets;
 
+        /// <summary>
+        /// Android Xamarin implementation of <see cref="ITextFileLoader"/>
+        /// </summary>
+        /// <param name="assets"></param>
         public AndroidTextFileImporter(AssetManager assets)
         {
-            _assets = assets ?? throw new System.ArgumentNullException(nameof(assets));
+            _assets = assets ?? throw new ArgumentNullException(nameof(assets));
         }
 
+        /// <summary>
+        /// Loads a text file from path
+        /// </summary>
+        /// <param name="filePath"></param>
+        /// <returns></returns>
         public string LoadFile(string filePath)
         {
             using (var fileStream = _assets.Open(filePath))
