+++ FbonizziMonoGameAndroid/FbonizziMonoGameAndroid/packages.config
@@ -1,18 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="MonoGame.Framework.Android" version="3.6.0.1625" targetFramework="monoandroid71" />
-  <package id="Xamarin.Android.Support.Compat" version="25.1.0" targetFramework="monoandroid71" />
-  <package id="Xamarin.Android.Support.Core.UI" version="25.1.0" targetFramework="monoandroid71" />
-  <package id="Xamarin.Android.Support.Core.Utils" version="25.1.0" targetFramework="monoandroid71" />
-  <package id="Xamarin.Android.Support.Fragment" version="25.1.0" targetFramework="monoandroid71" />
-  <package id="Xamarin.Android.Support.Media.Compat" version="25.1.0" targetFramework="monoandroid71" />
-  <package id="Xamarin.Android.Support.v4" version="25.1.0" targetFramework="monoandroid71" />
-  <package id="Xamarin.Build.Download" version="0.4.2" targetFramework="monoandroid71" />
-  <package id="Xamarin.GooglePlayServices.Ads" version="42.1021.1" targetFramework="monoandroid71" />
-  <package id="Xamarin.GooglePlayServices.Ads.Lite" version="42.1021.1" targetFramework="monoandroid71" />
-  <package id="Xamarin.GooglePlayServices.Base" version="42.1021.1" targetFramework="monoandroid71" />
-  <package id="Xamarin.GooglePlayServices.Basement" version="42.1021.1" targetFramework="monoandroid71" />
-  <package id="Xamarin.GooglePlayServices.Clearcut" version="42.1021.1" targetFramework="monoandroid71" />
-  <package id="Xamarin.GooglePlayServices.Gass" version="42.1021.1" targetFramework="monoandroid71" />
-  <package id="Xamarin.GooglePlayServices.Tasks" version="42.1021.1" targetFramework="monoandroid71" />
+  <package id="MonoGame.Framework.Android" version="3.7.0.1708" targetFramework="monoandroid81" />
 </packages>
