+++ FbonizziMonoGameUWP/FbonizziMonoGameUWP.nuspec
@@ -1,22 +1,25 @@
 <?xml version="1.0" encoding="utf-8"?>
 <package xmlns="http://schemas.microsoft.com/packaging/2010/07/nuspec.xsd">
   <metadata>
-    <id>FlowsoftGamesUWP</id>
-    <version>$version$</version>
-    <title>FlowsoftGamesUWP</title>
-    <authors>Flowsoft Games</authors>
-    <owners>Flowsoft Games</owners>
+    <id>FbonizziMonoGame.UWP</id>
+    <version>1.0.0-alpha</version>
+    <title>FbonizziMonoGame.UWP</title>
+    <authors>Francesco Bonizzi</authors>
+    <owners>Francesco Bonizzi</owners>
     <description>FbonizziMonoGame UWP implementation</description>
     <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <tags>monogame flowsoft games UWP</tags>
+    <tags>monogame francesco bonizzi games videogames UWP UAP</tags>
+    
     <dependencies>
       <group targetFramework="UAP10"> 
-        <dependency id="FbonizziMonoGame" version="2018.1.10.2" />
+        <dependency id="FbonizziMonoGame" version="1.0.0-alpha" />
+        <dependency id="Microsoft.NETCore.UniversalWindowsPlatform" version="6.2.0" />
       </group>
     </dependencies>
   </metadata>
+  
   <files>
-    <file src="FlowsoftGamesUWP\bin\Release\FlowsoftGamesUWP.dll" target="lib\UAP10\FlowsoftGamesUWP.dll" />
-    <file src="FlowsoftGamesUWP\bin\Release\FlowsoftGamesUWP.xml" target="lib\UAP10\FlowsoftGamesUWP.xml" />
+    <file src="\FbonizziMonoGameUWP\bin\Release\FlowsoftGamesUWP.*" target="lib\uap10" />
   </files>
+  
 </package>
