+++ FbonizziMonoGameWindowsDesktop/FbonizziMonoGameWindowsDesktop/FileWindowsSettingsRepository.cs
@@ -1,17 +1,29 @@
-﻿using FbonizziMonogame;
+﻿using FbonizziMonoGame.PlatformAbstractions;
 using System;
 using System.Collections.Generic;
 using System.IO;
 
 namespace FbonizziMonoGameWindowsDesktop
 {
-    public class WindowsSettingsRepository : ISettingsRepository
+    /// <summary>
+    /// An <see cref="ISettingsRepository"/> that saves a file in the format 'key:value\n'
+    /// </summary>
+    public class FileWindowsSettingsRepository : ISettingsRepository
     {
-        private const string _fileName = "gameSettings.txt";
+        private readonly string _fileName;
         private Dictionary<string, string> _storage;
 
-        public WindowsSettingsRepository()
+        /// <summary>
+        /// An <see cref="ISettingsRepository"/> that saves a file in the format 'key:value\n'
+        /// </summary>
+        /// <param name="settingsFilePath">The settings file path</param>
+        public FileWindowsSettingsRepository(string settingsFilePath)
         {
+            if (string.IsNullOrWhiteSpace(settingsFilePath))
+                throw new ArgumentNullException(nameof(settingsFilePath));
+
+            _fileName = settingsFilePath;
+
             if (File.Exists(_fileName))
             {
                 Deserialize(File.ReadAllLines(_fileName));
@@ -51,6 +63,12 @@ namespace FbonizziMonoGameWindowsDesktop
         private void Save()
             => File.WriteAllText(_fileName, Serialize());
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public bool GetOrSetBool(string key, bool defaultValue)
         {
             if (_storage.ContainsKey(key))
@@ -61,6 +79,12 @@ namespace FbonizziMonoGameWindowsDesktop
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public DateTime GetOrSetDateTime(string key, DateTime defaultValue)
         {
             if (_storage.ContainsKey(key))
@@ -71,6 +95,12 @@ namespace FbonizziMonoGameWindowsDesktop
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public int GetOrSetInt(string key, int defaultValue)
         {
             if (_storage.ContainsKey(key))
@@ -81,6 +111,12 @@ namespace FbonizziMonoGameWindowsDesktop
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public long GetOrSetLong(string key, long defaultValue)
         {
             if (_storage.ContainsKey(key))
@@ -91,6 +127,12 @@ namespace FbonizziMonoGameWindowsDesktop
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public string GetOrSetString(string key, string defaultValue)
         {
             if (_storage.ContainsKey(key))
@@ -101,6 +143,12 @@ namespace FbonizziMonoGameWindowsDesktop
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public TimeSpan GetOrSetTimeSpan(string key, TimeSpan defaultValue)
         {
             if (_storage.ContainsKey(key))
@@ -114,36 +162,67 @@ namespace FbonizziMonoGameWindowsDesktop
         private string FromBool(bool value)
             => value ? "1" : "0";
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetBool(string key, bool value)
         {
             _storage[key] = FromBool(value);
             Save();
         }
 
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetDateTime(string key, DateTime value)
         {
             SetDateTime(key, value);
             Save();
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetInt(string key, int value)
         {
             _storage[key] = value.ToString();
             Save();
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetLong(string key, long value)
         {
             _storage[key] = value.ToString();
             Save();
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetString(string key, string value)
         {
             _storage[key] = value;
             Save();
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetTimeSpan(string key, TimeSpan value)
         {
             SetLong(key, value.Ticks);
