+++ FbonizziMonoGameWindowsDesktop/FbonizziMonoGameWindowsDesktop/TransientWindowsSettingsRepository.cs
@@ -1,13 +1,22 @@
-﻿using FbonizziMonogame;
+﻿using FbonizziMonoGame.PlatformAbstractions;
 using System;
 using System.Collections.Generic;
 
 namespace FbonizziMonoGameWindowsDesktop
 {
+    /// <summary>
+    /// An in memory <see cref="ISettingsRepository"/> that is cleaned after the process ends
+    /// </summary>
     public class TransientWindowsSettingsRepository : ISettingsRepository
     {
         private Dictionary<string, object> _storage = new Dictionary<string, object>();
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public bool GetOrSetBool(string key, bool defaultValue)
         {
             if (_storage.ContainsKey(key))
@@ -17,6 +26,12 @@ namespace FbonizziMonoGameWindowsDesktop
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public int GetOrSetInt(string key, int defaultValue)
         {
             if (_storage.ContainsKey(key))
@@ -26,6 +41,12 @@ namespace FbonizziMonoGameWindowsDesktop
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public long GetOrSetLong(string key, long defaultValue)
         {
             if (_storage.ContainsKey(key))
@@ -35,6 +56,12 @@ namespace FbonizziMonoGameWindowsDesktop
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public string GetOrSetString(string key, string defaultValue)
         {
             if (_storage.ContainsKey(key))
@@ -44,6 +71,12 @@ namespace FbonizziMonoGameWindowsDesktop
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public TimeSpan GetOrSetTimeSpan(string key, TimeSpan defaultValue)
         {
             if (_storage.ContainsKey(key))
@@ -53,6 +86,12 @@ namespace FbonizziMonoGameWindowsDesktop
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="defaultValue"></param>
+        /// <returns></returns>
         public DateTime GetOrSetDateTime(string key, DateTime defaultValue)
         {
             if (_storage.ContainsKey(key))
@@ -62,6 +101,11 @@ namespace FbonizziMonoGameWindowsDesktop
             return defaultValue;
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetBool(string key, bool value)
         {
             if (_storage.ContainsKey(key))
@@ -70,6 +114,11 @@ namespace FbonizziMonoGameWindowsDesktop
                 _storage.Add(key, value);
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetInt(string key, int value)
         {
             if (_storage.ContainsKey(key))
@@ -78,6 +127,11 @@ namespace FbonizziMonoGameWindowsDesktop
                 _storage.Add(key, value);
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetLong(string key, long value)
         {
             if (_storage.ContainsKey(key))
@@ -86,6 +140,11 @@ namespace FbonizziMonoGameWindowsDesktop
                 _storage.Add(key, value);
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetString(string key, string value)
         {
             if (_storage.ContainsKey(key))
@@ -94,6 +153,11 @@ namespace FbonizziMonoGameWindowsDesktop
                 _storage.Add(key, value);
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetTimeSpan(string key, TimeSpan value)
         {
             if (_storage.ContainsKey(key))
@@ -102,6 +166,11 @@ namespace FbonizziMonoGameWindowsDesktop
                 _storage.Add(key, value);
         }
 
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="key"></param>
+        /// <param name="value"></param>
         public void SetDateTime(string key, DateTime value)
         {
             if (_storage.ContainsKey(key))
